Extract default transformers to their own file
