+++ .changeset/nervous-mangos-rhyme.md
@@ -1,0 +1,5 @@
+---
+'@backstage/plugin-catalog-backend-module-msgraph': patch
+---
+
+Extract default transformers to their own file
